Add files via upload
- datasetprocessed.xlsx
- datapreprocessing.ipynb

both files uploaded show the code used to fill in missing values in the dataset. the results are shown in the Excel file.
Some variables have also been removed since it does not correlate with the goal of the project.
the dataset has also been formatted into the same numerical format to ensure consistency.